Applied K-NN Classification Model and Evaluated Accordingly
Added a K-NN Classification Model to see if the following variables can be reliable predictors for whether someone will subscribe or not. Then, I evaluated whether or not it was accurate or not using accuracy and cross-validation. Additionally, I added library(themis) and step_upsample to balance the imbalanced data.